refactor: rename the search, video and playlist types
